+++ src/main/java/com/planbase/pdf/layoutmanager/contents/TableRow.kt
@@ -100,28 +100,31 @@ class TableRow(private val tablePart: TablePart) {
                 row.cells.map { c -> c!!.wrap() }
                         .toList()
 
-        fun render(lp: RenderTarget, topLeft: Coord): Dim {
+        fun render(lp: RenderTarget, topLeft: Coord, reallyRender: Boolean): Dim {
 //        cells.map { c -> c?.wrap() ?: LineWrapped.ZeroLineWrapped }
 //                .forEach{ c -> minRowHeight = Math.max(minRowHeight, c.dim.height)}
 
             var x = topLeft.x
             var maxRowHeight = minRowHeight
-//            println("minRowHeight=$minRowHeight")
+//            println("    minRowHeight=$minRowHeight")
             // Find the height of the tallest cell before rendering any cells.
             for (fixedCell in fixedCells) {
+//                println("    beforeRender height=${fixedCell.dim.height}")
                 val (width, height) = fixedCell.tableRender(lp, topLeft.x(x), maxRowHeight, false)
-//                println("height=$height")
+//                println("    afterRender height=$height") // Size is wrong here!
                 maxRowHeight = max(maxRowHeight, height)
-//                println("maxRowHeight=$maxRowHeight")
+//                println("    maxRowHeight=$maxRowHeight")
                 x += width
             }
             val maxWidth = x - topLeft.x
 
-            // Now render the cells
-            x = topLeft.x
-            for (fixedCell in fixedCells) {
-                val (width, _) = fixedCell.tableRender(lp, topLeft.x(x), maxRowHeight, true)
-                x += width
+            if (reallyRender) {
+                // Now render the cells
+                x = topLeft.x
+                for (fixedCell in fixedCells) {
+                    val (width, _) = fixedCell.tableRender(lp, topLeft.x(x), maxRowHeight, true)
+                    x += width
+                }
             }
 
             return Dim(maxWidth, maxRowHeight)