+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.3 - 2018-01-09
+ - Fixed: Line-wrapping sometimes drops the space before the last word on a line.
+   Only happened when the last word should have been too long to line wrap by a width of less than one space.
+ - Fixed: When nesting tables, the inner table gets extra space beneath it after a page break.
+   Went back to earlier idea of adding reallyRender boolean parameter to LineWrapped.render() so
+   that we could share all the logic between measuring and actually rendering.
+
 ## 2.0.2 - 2018-01-01
  - Upgraded Apache PDFBox to 2.0.0 and Kotlin to 1.2.10
  - Added this changelog